Renamed departmentReducers to departmentReducer
