+++ src/reducers/index.js
@@ -1,6 +1,6 @@
 import { combineReducers } from "redux";
 import authReducer from "./auth/authReducers";
-import departmentReducers from "./department/departmentReducers";
+import departmentReducers from "./department/departmentReducer";
 import jobsReducers from "./jobs/jobsReducer";
 
 export default combineReducers({
